+++ src/audio/class_audio.cpp
@@ -426,7 +426,7 @@ static ERR AUDIO_Beep(extAudio *Self, struct snd::Beep *Args)
 
 #ifdef __linux__
    if (auto console = GetResource(RES::CONSOLE_FD); console != -1) {
-      ioctl(console, KDMKTONE, ((1193190 / Args->Pitch) & 0xffff) | ((ULONG)Args->Duration << 16));
+      ioctl(console, KDMKTONE, ((1193190 / Args->Pitch) & 0xffff) | ((uint32_t)Args->Duration << 16));
       return ERR::Okay;
    }
 #elif _WIN32
