+++ src/core/tests/object_locking.cpp
@@ -18,8 +18,8 @@ using namespace pf;
 
 CSTRING ProgName = "ObjectLocking";
 static volatile OBJECTPTR glConfig = nullptr;
-static ULONG glTotalThreads = 8;
-static ULONG glLockAttempts = 200;
+static uint32_t glTotalThreads = 8;
+static uint32_t glLockAttempts = 200;
 static LONG glAccessGap = 200000;
 static bool glTerminateObject = false;
