+++ src/display/android/lib_pixels.cpp
@@ -2,7 +2,7 @@
 ** CHUNKY32
 */
 
-static void VideoDrawPixel32(objBitmap *Bitmap, LONG X, LONG Y, ULONG Colour)
+static void VideoDrawPixel32(objBitmap *Bitmap, LONG X, LONG Y, uint32_t Colour)
 {
 }
 
@@ -10,12 +10,12 @@ static void VideoDrawRGBPixel32(objBitmap *Bitmap, LONG X, LONG Y, struct RGB8 *
 {
 }
 
-static void VideoDrawRGBIndex32(objBitmap *Bitmap, ULONG *Data, struct RGB8 *RGB)
+static void VideoDrawRGBIndex32(objBitmap *Bitmap, uint32_t *Data, struct RGB8 *RGB)
 {
 
 }
 
-static ULONG VideoReadPixel32(objBitmap *Bitmap, LONG X, LONG Y)
+static uint32_t VideoReadPixel32(objBitmap *Bitmap, LONG X, LONG Y)
 {
    return 0;
 }
@@ -24,7 +24,7 @@ static void VideoReadRGBPixel32(objBitmap *Bitmap, LONG X, LONG Y, struct RGB8 *
 {
 }
 
-static void VideoReadRGBIndex32(objBitmap *Bitmap, ULONG *Data, struct RGB8 *RGB)
+static void VideoReadRGBIndex32(objBitmap *Bitmap, uint32_t *Data, struct RGB8 *RGB)
 {
 }
 
@@ -32,7 +32,7 @@ static void VideoReadRGBIndex32(objBitmap *Bitmap, ULONG *Data, struct RGB8 *RGB
 ** CHUNKY24
 */
 
-static void VideoDrawPixel24(objBitmap *Bitmap, LONG X, LONG Y, ULONG Colour)
+static void VideoDrawPixel24(objBitmap *Bitmap, LONG X, LONG Y, uint32_t Colour)
 {
 }
 
@@ -45,7 +45,7 @@ static void VideoDrawRGBIndex24(objBitmap *Bitmap, UBYTE *Data, struct RGB8 *RGB
 
 }
 
-static ULONG VideoReadPixel24(objBitmap *Bitmap, LONG X, LONG Y)
+static uint32_t VideoReadPixel24(objBitmap *Bitmap, LONG X, LONG Y)
 {
    return 0;
 }
@@ -63,7 +63,7 @@ static void VideoReadRGBIndex24(objBitmap *Bitmap, UBYTE *Data, struct RGB8 *RGB
 ** CHUNKY16
 */
 
-static void VideoDrawPixel16(objBitmap *Bitmap, LONG X, LONG Y, ULONG Colour)
+static void VideoDrawPixel16(objBitmap *Bitmap, LONG X, LONG Y, uint32_t Colour)
 {
 }
 
@@ -76,7 +76,7 @@ static void VideoDrawRGBIndex16(objBitmap *Bitmap, uint16_t *Data, struct RGB*RG
 
 }
 
-static ULONG VideoReadPixel16(objBitmap *Bitmap, LONG X, LONG Y)
+static uint32_t VideoReadPixel16(objBitmap *Bitmap, LONG X, LONG Y)
 {
    return 0;
 }
@@ -94,7 +94,7 @@ static void VideoReadRGBIndex16(objBitmap *Bitmap, uint16_t *Data, struct RGB*RG
 ** CHUNKY8
 */
 
-static void VideoDrawPixel8(objBitmap *Bitmap, LONG X, LONG Y, ULONG Colour)
+static void VideoDrawPixel8(objBitmap *Bitmap, LONG X, LONG Y, uint32_t Colour)
 {
 }
 
@@ -107,7 +107,7 @@ static void VideoDrawRGBIndex8(objBitmap *Bitmap, UBYTE *Data, struct RGB8 *RGB)
 
 }
 
-static ULONG VideoReadPixel8(objBitmap *Bitmap, LONG X, LONG Y)
+static uint32_t VideoReadPixel8(objBitmap *Bitmap, LONG X, LONG Y)
 {
    return 0;
 }
